added if/else for already parsed data. Fixed syllable splitting function for strings without alphabetic characters, e.g. "* * *"

+++ parser_files.py
@@ -12,7 +12,6 @@ print('list of texts')
 print(texts)
 # check folders
 new_folder = "text_ready/"
-print("sono qui1")
 try:
     os.stat(new_folder)
 except:
@@ -40,35 +39,38 @@ regola_char = re.compile(
 regola_space = re.compile("[’\']")
 
 for f in texts:
-    print(path+f)
-    text = open(path+f, encoding="utf8")
-    lines = text.readlines()
-    text.close()
-    # work on the single file
-    new_file = open(new_folder+"ready_"+f, "w")
-    i = False  # activator
-    for l in lines:
-        if 22 > len(l) or len(l) > 60:
-            continue
-        elif l[0] == "↑":
-            continue
-        else:
-            line=regola_char.sub("", 
-                           regola_eacc.sub("è",
-                           regola_i.sub("i", 
-                           regola_a.sub('a', 
-                           regola_space.sub(" ", 
-                           regola_e.sub("e", 
-                           regola_o.sub("o", 
-                           regola_oacc.sub("ó",
-                           regola_u.sub("u", 
-                           regola_iacc.sub("ì", 
-                           regola_aacc.sub("à", 
-                           regola_uacc.sub("ù", l)))))))))))).lower()
-            print(line)
-            print(sf.syllable_division(line))
-            print(sf.count_syllable(line))
-            print("________________________")
-            if 10 < sf.count_syllable(line) < 15: #TODO check if this interval 
-                new_file.write(line)
-    new_file.close()
+    if os.path.isfile(os.path.join(new_folder, f'ready_{f}')):
+        print(f'File {f} already parsed')
+    else:
+        print(path+f)
+        text = open(path+f, encoding="utf8")
+        lines = text.readlines()
+        text.close()
+        # work on the single file
+        new_file = open(new_folder+"ready_"+f, "w")
+        i = False  # activator
+        for l in lines:
+            if len(l) < 22 or len(l) > 60:
+                continue
+            elif l[0] == "↑":
+                continue
+            else:
+                line=regola_char.sub("",
+                               regola_eacc.sub("è",
+                               regola_i.sub("i",
+                               regola_a.sub('a',
+                               regola_space.sub(" ",
+                               regola_e.sub("e",
+                               regola_o.sub("o",
+                               regola_oacc.sub("ó",
+                               regola_u.sub("u",
+                               regola_iacc.sub("ì",
+                               regola_aacc.sub("à",
+                               regola_uacc.sub("ù", l)))))))))))).lower()
+                print(line)
+                print(sf.syllable_division(line))
+                print(sf.count_syllable(line))
+                print("________________________")
+                if 10 < sf.count_syllable(line) < 15: #TODO check if this interval
+                    new_file.write(line)
+        new_file.close()
